Extract bundle manifest creation into an ICommand.
