+++ src/tools/wix/Bind/BindBundleCommand.cs
@@ -1237,591 +1237,23 @@ namespace WixToolset.Bind
 
         private void CreateBurnManifest(string outputPath, WixBundleRow bundleInfo, WixBundleUpdateRow updateRow, WixUpdateRegistrationRow updateRegistrationInfo, string path, List<RelatedBundleInfo> allRelatedBundles, List<VariableInfo> allVariables, List<WixSearchInfo> orderedSearches, Dictionary<string, PayloadInfoRow> allPayloads, ChainInfo chain, Dictionary<string, ContainerInfo> containers, IEnumerable<WixCatalogRow> catalogs, Table wixBundleTagTable, IEnumerable<WixApprovedExeForElevationRow> approvedExesForElevation)
         {
-            string executableName = Path.GetFileName(outputPath);
-
-            using (XmlTextWriter writer = new XmlTextWriter(path, Encoding.UTF8))
-            {
-                writer.WriteStartDocument();
-
-                writer.WriteStartElement("BurnManifest", BurnCommon.BurnNamespace);
-
-                // Write the condition, if there is one
-                if (null != bundleInfo.Condition)
-                {
-                    writer.WriteElementString("Condition", bundleInfo.Condition);
-                }
-
-                // Write the log element if default logging wasn't disabled.
-                if (!String.IsNullOrEmpty(bundleInfo.LogPrefix))
-                {
-                    writer.WriteStartElement("Log");
-                    if (!String.IsNullOrEmpty(bundleInfo.LogPathVariable))
-                    {
-                        writer.WriteAttributeString("PathVariable", bundleInfo.LogPathVariable);
-                    }
-                    writer.WriteAttributeString("Prefix", bundleInfo.LogPrefix);
-                    writer.WriteAttributeString("Extension", bundleInfo.LogExtension);
-                    writer.WriteEndElement();
-                }
-
-                if (null != updateRow)
-                {
-                    writer.WriteStartElement("Update");
-                    writer.WriteAttributeString("Location", updateRow.Location);
-                    writer.WriteEndElement(); // </Update>
-                }
-
-                // Write the RelatedBundle elements
-                foreach (RelatedBundleInfo relatedBundle in allRelatedBundles)
-                {
-                    relatedBundle.WriteXml(writer);
-                }
-
-                // Write the variables
-                foreach (VariableInfo variable in allVariables)
-                {
-                    variable.WriteXml(writer);
-                }
-
-                // Write the searches
-                foreach (WixSearchInfo searchinfo in orderedSearches)
-                {
-                    searchinfo.WriteXml(writer);
-                }
-
-                // write the UX element
-                writer.WriteStartElement("UX");
-                if (!String.IsNullOrEmpty(bundleInfo.SplashScreenBitmapPath))
-                {
-                    writer.WriteAttributeString("SplashScreen", "yes");
-                }
-
-                // write the UX allPayloads...
-                List<PayloadInfoRow> uxPayloads = containers[Compiler.BurnUXContainerId].Payloads;
-                foreach (PayloadInfoRow payload in uxPayloads)
-                {
-                    writer.WriteStartElement("Payload");
-                    WriteBurnManifestPayloadAttributes(writer, payload, true, allPayloads);
-                    writer.WriteEndElement();
-                }
-
-                writer.WriteEndElement();
-
-                // write the catalog elements
-                if (catalogs.Any())
-                {
-                    foreach (WixCatalogRow catalog in catalogs)
-                    {
-                        writer.WriteStartElement("Catalog");
-                        writer.WriteAttributeString("Id", catalog.Id);
-                        writer.WriteAttributeString("Payload", catalog.PayloadId);
-                        writer.WriteEndElement();
-                    }
-                }
-
-                int attachedContainerIndex = 1; // count starts at one because UX container is "0".
-                foreach (ContainerInfo container in containers.Values)
-                {
-                    if (Compiler.BurnUXContainerId != container.Id && 0 < container.Payloads.Count)
-                    {
-                        writer.WriteStartElement("Container");
-                        WriteBurnManifestContainerAttributes(writer, executableName, container, attachedContainerIndex);
-                        writer.WriteEndElement();
-                        if ("attached" == container.Type)
-                        {
-                            attachedContainerIndex++;
-                        }
-                    }
-                }
-
-                foreach (PayloadInfoRow payload in allPayloads.Values)
-                {
-                    if (PackagingType.Embedded == payload.Packaging && Compiler.BurnUXContainerId != payload.Container)
-                    {
-                        writer.WriteStartElement("Payload");
-                        WriteBurnManifestPayloadAttributes(writer, payload, true, allPayloads);
-                        writer.WriteEndElement();
-                    }
-                    else if (PackagingType.External == payload.Packaging)
-                    {
-                        writer.WriteStartElement("Payload");
-                        WriteBurnManifestPayloadAttributes(writer, payload, false, allPayloads);
-                        writer.WriteEndElement();
-                    }
-                }
-
-                foreach (RollbackBoundaryInfo rollbackBoundary in chain.RollbackBoundaries)
-                {
-                    writer.WriteStartElement("RollbackBoundary");
-                    writer.WriteAttributeString("Id", rollbackBoundary.Id);
-                    writer.WriteAttributeString("Vital", YesNoType.Yes == rollbackBoundary.Vital ? "yes" : "no");
-                    writer.WriteEndElement();
-                }
-
-                // Write the registration information...
-                writer.WriteStartElement("Registration");
-
-                writer.WriteAttributeString("Id", bundleInfo.BundleId.ToString("B"));
-                writer.WriteAttributeString("ExecutableName", executableName);
-                writer.WriteAttributeString("PerMachine", bundleInfo.PerMachine ? "yes" : "no");
-                writer.WriteAttributeString("Tag", bundleInfo.Tag);
-                writer.WriteAttributeString("Version", bundleInfo.Version);
-                writer.WriteAttributeString("ProviderKey", bundleInfo.ProviderKey);
-
-                writer.WriteStartElement("Arp");
-                writer.WriteAttributeString("Register", (0 < bundleInfo.DisableModify && bundleInfo.DisableRemove) ? "no" : "yes"); // do not register if disabled modify and remove.
-                writer.WriteAttributeString("DisplayName", bundleInfo.Name);
-                writer.WriteAttributeString("DisplayVersion", bundleInfo.Version);
-
-                if (!String.IsNullOrEmpty(bundleInfo.Publisher))
-                {
-                    writer.WriteAttributeString("Publisher", bundleInfo.Publisher);
-                }
-
-                if (!String.IsNullOrEmpty(bundleInfo.HelpLink))
-                {
-                    writer.WriteAttributeString("HelpLink", bundleInfo.HelpLink);
-                }
-
-                if (!String.IsNullOrEmpty(bundleInfo.HelpTelephone))
-                {
-                    writer.WriteAttributeString("HelpTelephone", bundleInfo.HelpTelephone);
-                }
-
-                if (!String.IsNullOrEmpty(bundleInfo.AboutUrl))
-                {
-                    writer.WriteAttributeString("AboutUrl", bundleInfo.AboutUrl);
-                }
-
-                if (!String.IsNullOrEmpty(bundleInfo.UpdateUrl))
-                {
-                    writer.WriteAttributeString("UpdateUrl", bundleInfo.UpdateUrl);
-                }
-
-                if (!String.IsNullOrEmpty(bundleInfo.ParentName))
-                {
-                    writer.WriteAttributeString("ParentDisplayName", bundleInfo.ParentName);
-                }
-
-                if (1 == bundleInfo.DisableModify)
-                {
-                    writer.WriteAttributeString("DisableModify", "yes");
-                }
-                else if (2 == bundleInfo.DisableModify)
-                {
-                    writer.WriteAttributeString("DisableModify", "button");
-                }
-
-                if (bundleInfo.DisableRemove)
-                {
-                    writer.WriteAttributeString("DisableRemove", "yes");
-                }
-                writer.WriteEndElement(); // </Arp>
-
-                if (null != updateRegistrationInfo)
-                {
-                    writer.WriteStartElement("Update"); // <Update>
-                    writer.WriteAttributeString("Manufacturer", updateRegistrationInfo.Manufacturer);
-
-                    if (!String.IsNullOrEmpty(updateRegistrationInfo.Department))
-                    {
-                        writer.WriteAttributeString("Department", updateRegistrationInfo.Department);
-                    }
-
-                    if (!String.IsNullOrEmpty(updateRegistrationInfo.ProductFamily))
-                    {
-                        writer.WriteAttributeString("ProductFamily", updateRegistrationInfo.ProductFamily);
-                    }
-
-                    writer.WriteAttributeString("Name", updateRegistrationInfo.Name);
-                    writer.WriteAttributeString("Classification", updateRegistrationInfo.Classification);
-                    writer.WriteEndElement(); // </Update>
-                }
-
-                if (null != wixBundleTagTable)
-                {
-                    foreach (Row row in wixBundleTagTable.Rows)
-                    {
-                        writer.WriteStartElement("SoftwareTag");
-                        writer.WriteAttributeString("Filename", (string)row[0]);
-                        writer.WriteAttributeString("Regid", (string)row[1]);
-                        writer.WriteCData((string)row[4]);
-                        writer.WriteEndElement();
-                    }
-                }
-
-                writer.WriteEndElement(); // </Register>
-
-                // write the Chain...
-                writer.WriteStartElement("Chain");
-                if (chain.DisableRollback)
-                {
-                    writer.WriteAttributeString("DisableRollback", "yes");
-                }
-
-                if (chain.DisableSystemRestore)
-                {
-                    writer.WriteAttributeString("DisableSystemRestore", "yes");
-                }
-
-                if (chain.ParallelCache)
-                {
-                    writer.WriteAttributeString("ParallelCache", "yes");
-                }
-
-                // Build up the list of target codes from all the MSPs in the chain.
-                List<WixBundlePatchTargetCodeRow> targetCodes = new List<WixBundlePatchTargetCodeRow>();
-
-                foreach (ChainPackageInfo package in chain.Packages)
-                {
-                    writer.WriteStartElement(String.Format(CultureInfo.InvariantCulture, "{0}Package", package.ChainPackageType));
-
-                    writer.WriteAttributeString("Id", package.Id);
-
-                    switch (package.Cache)
-                    {
-                        case YesNoAlwaysType.No:
-                            writer.WriteAttributeString("Cache", "no");
-                            break;
-                        case YesNoAlwaysType.Yes:
-                            writer.WriteAttributeString("Cache", "yes");
-                            break;
-                        case YesNoAlwaysType.Always:
-                            writer.WriteAttributeString("Cache", "always");
-                            break;
-                    }
-
-                    writer.WriteAttributeString("CacheId", package.CacheId);
-                    writer.WriteAttributeString("InstallSize", Convert.ToString(package.InstallSize));
-                    writer.WriteAttributeString("Size", Convert.ToString(package.Size));
-                    writer.WriteAttributeString("PerMachine", YesNoDefaultType.Yes == package.PerMachine ? "yes" : "no");
-                    writer.WriteAttributeString("Permanent", package.Permanent ? "yes" : "no");
-                    writer.WriteAttributeString("Vital", package.Vital ? "yes" : "no");
-
-                    if (null != package.RollbackBoundary)
-                    {
-                        writer.WriteAttributeString("RollbackBoundaryForward", package.RollbackBoundary.Id);
-                    }
-
-                    if (!String.IsNullOrEmpty(package.RollbackBoundaryBackwardId))
-                    {
-                        writer.WriteAttributeString("RollbackBoundaryBackward", package.RollbackBoundaryBackwardId);
-                    }
-
-                    if (!String.IsNullOrEmpty(package.LogPathVariable))
-                    {
-                        writer.WriteAttributeString("LogPathVariable", package.LogPathVariable);
-                    }
-
-                    if (!String.IsNullOrEmpty(package.RollbackLogPathVariable))
-                    {
-                        writer.WriteAttributeString("RollbackLogPathVariable", package.RollbackLogPathVariable);
-                    }
-
-                    if (!String.IsNullOrEmpty(package.InstallCondition))
-                    {
-                        writer.WriteAttributeString("InstallCondition", package.InstallCondition);
-                    }
-
-                    if (Compiler.ChainPackageType.Exe == package.ChainPackageType)
-                    {
-                        writer.WriteAttributeString("DetectCondition", package.DetectCondition);
-                        writer.WriteAttributeString("InstallArguments", package.InstallCommand);
-                        writer.WriteAttributeString("UninstallArguments", package.UninstallCommand);
-                        writer.WriteAttributeString("RepairArguments", package.RepairCommand);
-                        writer.WriteAttributeString("Repairable", package.Repairable ? "yes" : "no");
-                        if (!String.IsNullOrEmpty(package.Protocol))
-                        {
-                            writer.WriteAttributeString("Protocol", package.Protocol);
-                        }
-                    }
-                    else if (Compiler.ChainPackageType.Msi == package.ChainPackageType)
-                    {
-                        writer.WriteAttributeString("ProductCode", package.ProductCode);
-                        writer.WriteAttributeString("Language", package.Language);
-                        writer.WriteAttributeString("Version", package.Version);
-                        writer.WriteAttributeString("DisplayInternalUI", package.DisplayInternalUI ? "yes" : "no");
-                        if (!String.IsNullOrEmpty(package.UpgradeCode))
-                        {
-                            writer.WriteAttributeString("UpgradeCode", package.UpgradeCode);
-                        }
-                    }
-                    else if (Compiler.ChainPackageType.Msp == package.ChainPackageType)
-                    {
-                        writer.WriteAttributeString("PatchCode", package.PatchCode);
-                        writer.WriteAttributeString("PatchXml", package.PatchXml);
-                        writer.WriteAttributeString("DisplayInternalUI", package.DisplayInternalUI ? "yes" : "no");
-
-                        // If there is still a chance that all of our patches will target a narrow set of
-                        // product codes, add the patch list to the overall list.
-                        if (null != targetCodes)
-                        {
-                            if (!package.TargetUnspecified)
-                            {
-                                targetCodes.AddRange(package.TargetCodes);
-                            }
-                            else // we have a patch that targets the world, so throw the whole list away.
-                            {
-                                targetCodes = null;
-                            }
-                        }
-                    }
-                    else if (Compiler.ChainPackageType.Msu == package.ChainPackageType)
-                    {
-                        writer.WriteAttributeString("DetectCondition", package.DetectCondition);
-                        writer.WriteAttributeString("KB", package.MsuKB);
-                    }
-
-                    foreach (MsiFeature feature in package.MsiFeatures)
-                    {
-                        writer.WriteStartElement("MsiFeature");
-                        writer.WriteAttributeString("Id", feature.Name);
-                        writer.WriteEndElement();
-                    }
-
-                    foreach (MsiPropertyInfo msiProperty in package.MsiProperties)
-                    {
-                        writer.WriteStartElement("MsiProperty");
-                        writer.WriteAttributeString("Id", msiProperty.Name);
-                        writer.WriteAttributeString("Value", msiProperty.Value);
-                        writer.WriteEndElement();
-                    }
-
-                    foreach (string slipstreamMsp in package.SlipstreamMsps)
-                    {
-                        writer.WriteStartElement("SlipstreamMsp");
-                        writer.WriteAttributeString("Id", slipstreamMsp);
-                        writer.WriteEndElement();
-                    }
-
-                    foreach (ExitCodeInfo exitCode in package.ExitCodes)
-                    {
-                        writer.WriteStartElement("ExitCode");
-                        writer.WriteAttributeString("Type", exitCode.Type);
-                        writer.WriteAttributeString("Code", exitCode.Code);
-                        writer.WriteEndElement();
-                    }
-
-                    // Output the dependency information.
-                    foreach (ProvidesDependency dependency in package.Provides)
-                    {
-                        // TODO: Add to wixpdb as an imported table, or link package wixpdbs to bundle wixpdbs.
-                        dependency.WriteXml(writer);
-                    }
-
-                    foreach (RelatedPackage related in package.RelatedPackages)
-                    {
-                        writer.WriteStartElement("RelatedPackage");
-                        writer.WriteAttributeString("Id", related.Id);
-                        if (!String.IsNullOrEmpty(related.MinVersion))
-                        {
-                            writer.WriteAttributeString("MinVersion", related.MinVersion);
-                            writer.WriteAttributeString("MinInclusive", related.MinInclusive ? "yes" : "no");
-                        }
-                        if (!String.IsNullOrEmpty(related.MaxVersion))
-                        {
-                            writer.WriteAttributeString("MaxVersion", related.MaxVersion);
-                            writer.WriteAttributeString("MaxInclusive", related.MaxInclusive ? "yes" : "no");
-                        }
-                        writer.WriteAttributeString("OnlyDetect", related.OnlyDetect ? "yes" : "no");
-                        if (0 < related.Languages.Count)
-                        {
-                            writer.WriteAttributeString("LangInclusive", related.LangInclusive ? "yes" : "no");
-                            foreach (string language in related.Languages)
-                            {
-                                writer.WriteStartElement("Language");
-                                writer.WriteAttributeString("Id", language);
-                                writer.WriteEndElement();
-                            }
-                        }
-                        writer.WriteEndElement();
-                    }
-
-                    // Write any contained Payloads with the PackagePayload being first
-                    writer.WriteStartElement("PayloadRef");
-                    writer.WriteAttributeString("Id", package.PackagePayload.Id);
-                    writer.WriteEndElement();
-
-                    foreach (PayloadInfoRow payload in package.Payloads)
-                    {
-                        if (payload.Id != package.PackagePayload.Id)
-                        {
-                            writer.WriteStartElement("PayloadRef");
-                            writer.WriteAttributeString("Id", payload.Id);
-                            writer.WriteEndElement();
-                        }
-                    }
-
-                    writer.WriteEndElement(); // </XxxPackage>
-                }
-                writer.WriteEndElement(); // </Chain>
-
-                if (null != targetCodes)
-                {
-                    foreach (WixBundlePatchTargetCodeRow targetCode in targetCodes)
-                    {
-                        writer.WriteStartElement("PatchTargetCode");
-                        writer.WriteAttributeString("TargetCode", targetCode.TargetCode);
-                        writer.WriteAttributeString("Product", targetCode.TargetsProductCode ? "yes" : "no");
-                        writer.WriteEndElement();
-                    }
-                }
-
-                // write the ApprovedExeForElevation elements
-                foreach (WixApprovedExeForElevationRow approvedExeForElevation in approvedExesForElevation)
-                {
-                    writer.WriteStartElement("ApprovedExeForElevation");
-                    writer.WriteAttributeString("Id", approvedExeForElevation.Id);
-                    writer.WriteAttributeString("Key", approvedExeForElevation.Key);
-
-                    if (!String.IsNullOrEmpty(approvedExeForElevation.ValueName))
-                    {
-                        writer.WriteAttributeString("ValueName", approvedExeForElevation.ValueName);
-                    }
-
-                    if (approvedExeForElevation.Win64)
-                    {
-                        writer.WriteAttributeString("Win64", "yes");
-                    }
-
-                    writer.WriteEndElement();
-                }
-
-                writer.WriteEndDocument(); // </BurnManifest>
-            }
-        }
-
-        private void WriteBurnManifestContainerAttributes(XmlTextWriter writer, string executableName, ContainerInfo container, int containerIndex)
-        {
-            writer.WriteAttributeString("Id", container.Id);
-            writer.WriteAttributeString("FileSize", container.FileInfo.Length.ToString(CultureInfo.InvariantCulture));
-            writer.WriteAttributeString("Hash", Common.GetFileHash(container.FileInfo));
-            if (container.Type == "detached")
-            {
-                string resolvedUrl = this.ResolveUrl(container.DownloadUrl, null, null, container.Id, container.Name);
-                if (!String.IsNullOrEmpty(resolvedUrl))
-                {
-                    writer.WriteAttributeString("DownloadUrl", resolvedUrl);
-                }
-                else if (!String.IsNullOrEmpty(container.DownloadUrl))
-                {
-                    writer.WriteAttributeString("DownloadUrl", container.DownloadUrl);
-                }
-
-                writer.WriteAttributeString("FilePath", container.Name);
-            }
-            else if (container.Type == "attached")
-            {
-                if (!String.IsNullOrEmpty(container.DownloadUrl))
-                {
-                    Messaging.Instance.OnMessage(WixWarnings.DownloadUrlNotSupportedForAttachedContainers(container.SourceLineNumbers, container.Id));
-                }
-
-                writer.WriteAttributeString("FilePath", executableName); // attached containers use the name of the bundle since they are attached to the executable.
-                writer.WriteAttributeString("AttachedIndex", containerIndex.ToString(CultureInfo.InvariantCulture));
-                writer.WriteAttributeString("Attached", "yes");
-                writer.WriteAttributeString("Primary", "yes");
-            }
-        }
-
-        private void WriteBurnManifestPayloadAttributes(XmlTextWriter writer, PayloadInfoRow payload, bool embeddedOnly, Dictionary<string, PayloadInfoRow> allPayloads)
-        {
-            Debug.Assert(!embeddedOnly || PackagingType.Embedded == payload.Packaging);
-
-            writer.WriteAttributeString("Id", payload.Id);
-            writer.WriteAttributeString("FilePath", payload.Name);
-            writer.WriteAttributeString("FileSize", payload.FileSize.ToString(CultureInfo.InvariantCulture));
-            writer.WriteAttributeString("Hash", payload.Hash);
-
-            if (payload.LayoutOnly)
-            {
-                writer.WriteAttributeString("LayoutOnly", "yes");
-            }
-
-            if (!String.IsNullOrEmpty(payload.PublicKey))
-            {
-                writer.WriteAttributeString("CertificateRootPublicKeyIdentifier", payload.PublicKey);
-            }
-
-            if (!String.IsNullOrEmpty(payload.Thumbprint))
-            {
-                writer.WriteAttributeString("CertificateRootThumbprint", payload.Thumbprint);
-            }
-
-            switch (payload.Packaging)
-            {
-                case PackagingType.Embedded: // this means it's in a container.
-                    if (!String.IsNullOrEmpty(payload.DownloadUrl))
-                    {
-                        Messaging.Instance.OnMessage(WixWarnings.DownloadUrlNotSupportedForEmbeddedPayloads(payload.SourceLineNumbers, payload.Id));
-                    }
-
-                    writer.WriteAttributeString("Packaging", "embedded");
-                    writer.WriteAttributeString("SourcePath", payload.EmbeddedId);
-
-                    if (Compiler.BurnUXContainerId != payload.Container)
-                    {
-                        writer.WriteAttributeString("Container", payload.Container);
-                    }
-                    break;
-
-                case PackagingType.External:
-                    string packageId = payload.ParentPackagePayload;
-                    string parentUrl = payload.ParentPackagePayload == null ? null : allPayloads[payload.ParentPackagePayload].DownloadUrl;
-                    string resolvedUrl = this.ResolveUrl(payload.DownloadUrl, parentUrl, packageId, payload.Id, payload.Name);
-                    if (!String.IsNullOrEmpty(resolvedUrl))
-                    {
-                        writer.WriteAttributeString("DownloadUrl", resolvedUrl);
-                    }
-                    else if (!String.IsNullOrEmpty(payload.DownloadUrl))
-                    {
-                        writer.WriteAttributeString("DownloadUrl", payload.DownloadUrl);
-                    }
-
-                    writer.WriteAttributeString("Packaging", "external");
-                    writer.WriteAttributeString("SourcePath", payload.Name);
-                    break;
-            }
-
-            if (!String.IsNullOrEmpty(payload.CatalogId))
-            {
-                writer.WriteAttributeString("Catalog", payload.CatalogId);
-            }
-        }
-
-        private string ResolveFile(string source, string type, SourceLineNumber sourceLineNumbers, BindStage bindStage = BindStage.Normal)
-        {
-            string path = null;
-            foreach (IBinderFileManager fileManager in this.FileManagers)
-            {
-                path = fileManager.ResolveFile(source, type, sourceLineNumbers, bindStage);
-                if (null != path)
-                {
-                    break;
-                }
-            }
-
-            if (null == path)
-            {
-                throw new WixFileNotFoundException(sourceLineNumbers, source, type);
-            }
-
-            return path;
-        }
-
-        private string ResolveUrl(string url, string fallbackUrl, string packageId, string payloadId, string fileName)
-        {
-            string resolved = null;
-            foreach (IBinderFileManager fileManager in this.FileManagers)
-            {
-                resolved = fileManager.ResolveUrl(url, fallbackUrl, packageId, payloadId, fileName);
-                if (!String.IsNullOrEmpty(resolved))
-                {
-                    break;
-                }
-            }
-
-            return resolved;
+            CreateBurnManifestCommand command = new CreateBurnManifestCommand();
+            command.FileManagers = this.FileManagers;
+            command.ExecutableName = Path.GetFileName(outputPath);
+            command.BundleInfo = bundleInfo;
+            command.UpdateRow = updateRow;
+            command.UpdateRegistrationInfo = updateRegistrationInfo;
+            command.OutputPath = path;
+            command.RelatedBundles = allRelatedBundles;
+            command.Variables = allVariables;
+            command.OrderedSearches = orderedSearches;
+            command.Payloads = allPayloads;
+            command.Chain = chain;
+            command.Containers = containers;
+            command.Catalogs = catalogs;
+            command.BundleTags = (null == wixBundleTagTable) ? Enumerable.Empty<Row>() : wixBundleTagTable.Rows;
+            command.ApprovedExesForElevation = approvedExesForElevation;
+            command.Execute();
         }
 
         #region DependencyExtension
